OCPBUGS-2992 Don't save OS_ prefixed variables
This change stops saving the OS_ prefixed variables to a file but keeps
them logged on stdout. The file is not consumed elsewhere, by either
this image or Ironic itself. Outputting to a container instance's
stdout allows some other log collector to save what may be helpful
debugging information about the Oslo configuration, and adding the:
`|| true` allows for base images without OS_ unset to cause a non-zero
return code with pipefail set.